pyproject.toml
[tool.poetry]
name = "ecommerce"
version = "0.1.0"
description = ""
authors = ["kirill <Richi.mosberg@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
fastapi = {extras = ["all"], version = "^0.109.0"}
uvicorn = "^0.27.0.post1"
email-validator = "^2.1.0.post1"
alembic = "^1.13.1"
black = "^24.1.1"
asyncpg = "^0.29.0"
bcrypt = "^4.1.2"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
pillow = "^10.2.0"
aioredis = "^2.0.1"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}
python-dotenv = "^1.0.1"
anyio = "^4.2.0"
pytest = "^7.3.2"
psycopg2 = "^2.9.9"
trio = "^0.24.0"
celery = {extras = ["redis"], version = "^5.3.6"}
image = "^1.5.33"
aiofiles = "^23.2.1"
python-json-logger = "^2.0.7"
aiohttp = "^3.9.3"
aiobotocore = "^2.13.0"
pytest-asyncio = "0.21.1"
sqlalchemy = "^2.0.30"
ruff = "^0.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# Set the maximum line length to 79.
line-length = 79

[tool.ruff.lint]
# Add the `line-too-long` rule to the enforced rule set.
extend-select = ["E501"]